requirements_dev.txt
ruff==0.13.2
mypy==1.18.2
pytest==8.4.2
pytest-asyncio==0.23.5
pytest-homeassistant-custom-component==0.13.109
cryptography==46.0.1
pre-commit==4.3.0
